+++ relay/kusama/src/governance/mod.rs
@@ -26,36 +26,11 @@ use frame_system::EnsureRootWithSuccess;
 mod origins;
 pub use origins::{
 	pallet_custom_origins, AuctionAdmin, Fellows, FellowshipAdmin, FellowshipExperts,
-	FellowshipInitiates, FellowshipMasters, GeneralAdmin, LeaseAdmin, Origin, ReferendumCanceller,
-	ReferendumKiller, Spender, StakingAdmin, Treasurer, WhitelistedCaller,
+	FellowshipInitiates, FellowshipMasters, GeneralAdmin, LeaseAdmin, Origin, Spender,
+	StakingAdmin, Treasurer, WhitelistedCaller,
 };
-mod tracks;
-pub use tracks::TracksInfo;
 mod fellowship;
 pub use fellowship::{FellowshipCollectiveInstance, FellowshipReferendaInstance};
-
-parameter_types! {
-	pub const VoteLockingPeriod: BlockNumber = 7 * DAYS;
-}
-
-impl pallet_conviction_voting::Config for Runtime {
-	type WeightInfo = weights::pallet_conviction_voting::WeightInfo<Self>;
-	type RuntimeEvent = RuntimeEvent;
-	type Currency = Balances;
-	type VoteLockingPeriod = VoteLockingPeriod;
-	type MaxVotes = ConstU32<512>;
-	type MaxTurnout =
-		frame_support::traits::tokens::currency::ActiveIssuanceOf<Balances, Self::AccountId>;
-	type Polls = Referenda;
-	type BlockNumberProvider = System;
-	type VotingHooks = ();
-}
-
-parameter_types! {
-	pub const AlarmInterval: BlockNumber = 1;
-	pub const SubmissionDeposit: Balance = 100 * QUID;
-	pub const UndecidingTimeout: BlockNumber = 14 * DAYS;
-}
 
 parameter_types! {
 	pub const MaxBalance: Balance = Balance::MAX;
@@ -72,25 +47,4 @@ impl pallet_whitelist::Config for Runtime {
 		EitherOf<EnsureRootWithSuccess<Self::AccountId, ConstU16<65535>>, Fellows>;
 	type DispatchWhitelistedOrigin = EitherOf<EnsureRoot<Self::AccountId>, WhitelistedCaller>;
 	type Preimages = Preimage;
-}
-
-impl pallet_referenda::Config for Runtime {
-	type WeightInfo = weights::pallet_referenda_referenda::WeightInfo<Self>;
-	type RuntimeCall = RuntimeCall;
-	type RuntimeEvent = RuntimeEvent;
-	type Scheduler = Scheduler;
-	type Currency = Balances;
-	type SubmitOrigin = frame_system::EnsureSigned<AccountId>;
-	type CancelOrigin = EitherOf<EnsureRoot<AccountId>, ReferendumCanceller>;
-	type KillOrigin = EitherOf<EnsureRoot<AccountId>, ReferendumKiller>;
-	type Slash = Treasury;
-	type Votes = pallet_conviction_voting::VotesOf<Runtime>;
-	type Tally = pallet_conviction_voting::TallyOf<Runtime>;
-	type SubmissionDeposit = SubmissionDeposit;
-	type MaxQueued = ConstU32<100>;
-	type UndecidingTimeout = UndecidingTimeout;
-	type AlarmInterval = AlarmInterval;
-	type Tracks = TracksInfo;
-	type Preimages = Preimage;
-	type BlockNumberProvider = System;
 }